[NCL-4253] unification of ArtifactImportError
